refactor: aztec l1 contracts deployed w/ forge (AztecProtocol/aztec-packages#18622)
Part of
https://linear.app/aztec-labs/issue/A-316/consolidate-contract-deployments

This PR refactors the deployment pipeline for governance and rollup
contracts. The typescript deployment mechanism is fully replaced with
Forge. The forge scripts take configuration using environment variables.
They write their deployed contract addresses to a JSON file, parsed by
the caller. The results are intended to be exactly the same, barring
minor things (e.g. deploying the mock zkpassport verifier in more
cases).

Core files added:
    - DeployAztecL1Contracts.s.sol - Full deployment (fresh network)
    - DeployRollupForUpgrade.s.sol - Upgrade-only deployment
    - DeployRollupLib.sol - Shared rollup deployment library
- DeploymentConfiguration.sol - Full deployment config. Reads env vars,
provides defaults based on network label
- RollupConfiguration.sol - Rollup-specific config (genesis state,
staking stuff, initial validators etc.)

Then lots of work making sure the right glue was in place for the
typescript side.

#
Why move to Forge for deployments?

L1 contract deployment was being done in TypeScript but clearly from the
network launching using foundry scripts, this was not ideal.

Benefits:
- forge --verify automatically posts to etherscan for us, avoiding the
complicated custom logic we had to implement
- forge simulation means we catch any problems in the entire deployment
flow early. Stuff like the previous validateConfig() that guarded
against reverts are now redundant.
- forge is able to broadcast these in transaction batches, I got 2-3
slots on Sepolia for a deploy (24-40s)
- better separation of concerns, l1-contracts now can fully own the
deployment
- foundry tests are very effective at asserting things about our
deployments under different conditions

Weirdness:
- we work around a problem in solidity where MockBlobLib was not
deployable through forge but it feels a bit hacky. I think it's
preferable to needing to rebuild with a different forge profile like
ignition-monorepo did.

Cut features:
- Deploying with a salt is no longer supported. While nice for
deterministic deployments, the reason the salt was useful before was
mainly to batch deployments. We do this with Forge, so forego the
complexity of supporting this.
This is feasible to bring back if it comes up, just would add a bunch of
conditionality to the script.

Comments on future work:
- It feels like the definition of networks is a bit sprawling and a bit
hard to analyze, as the env vars that trigger deployment can be set very
high up in the stack, in e.g. test scenario YAMLs. It would be best IMO
to make each value only defined in a specific module. Given our limited
set of distinct scenarios, we could lean in to the network label
approach like in previous ethereum/config.ts (and now
DeploymentConfiguration).

In that world, l1-contracts would choose the parameters based on a
network label and no other part of the stack could inject values thru
env vars, making analysis easier. This would be more like the
ignition-monorepo deploys (but they had fewer variations to contend
with).

For now it is a bit of a mix as network-specific governance
configuration definitions (local, staging, testnet, mainnet) from
TypeScript's config.ts have been moved to DeploymentConfiguration.sol
but it allows for a lot to be overridden with env variables.