fix: Disable json rpc batching in viem http transports (AztecProtocol/aztec-packages#19073)
We are getting sporadic `Transaction creation failed` in next. These are
produced by viem when the execution client returns a -32003 error code,
which geth produces only when the size of a batch response exceeds a
threshold.

Under the hood, viem defaults to batching requests. So if we disable
batching altogether, maybe we stop seeing this error. Worth a try.