refactor: more robust tagging index sync as recipient (AztecProtocol/aztec-packages#19030)
Fixes AztecProtocol/aztec-packages#17775

In this PR I implement a new log sync algorithm we've come up with in
Buenos Aires that should be fully robust against any kind of private log
losses. The algorithm is explained in the body of
`loadPrivateLogsForSenderRecipientPair` function. That function should
be the entrypoint when reviewing this PR (if checking the utility
functions first you would not have enough context).

Unfortunately this PR introduces a performance regression that is
tracked by [this
issue](https://linear.app/aztec-labs/issue/F-229/improve-log-sync-performance).
I am fairly certain that the regression is caused by an unreasonable
number of requests performed by the `loadLogsForRange` function - that
will be tackled in a followup PR.

In this PR the algorithm is not yet integrated into the system. That is
done in a [PR up the
stack](https://github.com/AztecProtocol/aztec-packages/pull/19125).

The directory structure is not yet final - I wanted to keep this PR
contained in one place to not have conflicts with Martin's PR. So please
ignore that for now (will move stuff around in a final pass).

My plan is as follows:
1. Merge this PR,
2. fix the regression by modifying the `getLogsByTag` API such that it
gives me all the info I need (now it doesn't give me block timestamp),
3. once that PR is merged most likely wait for Martin's refactor to be
merged and then rebase and polish my integrating new log sync PR,
4. move some files around to make it all cuter.

## Update on regression of time in e2e tests
Unfortunately realized that this has caused a regression and log sync is
now significantly slower.

Did 2 runs of the `e2e_l1_with_wall_time` test and on `next` the results
are 337 and 334 seconds but with the new code it's 661 and 658 seconds.

Will work on dropping that down before merging that "integrating new log
sync" PR.