refactor!: optimizing `AztecNode::getLogsByTags` for new log sync algo (AztecProtocol/aztec-packages#19157)
Fixes
https://linear.app/aztec-labs/issue/F-229/improve-log-sync-performance

The return type of `getLogsByTags` did not contain block timestamp which
lead to the new log sync algo being inefficient - we needed to perform a
lot of calls to `getBlockHeaderByHash` endpoint to get the timestamp.

In this PR I include the timestamp in the return value.

This is one of 3 PRs in which I clean up the `getLogsByTags` endpoint:
1. This is the first PR - including block timestamp,
2. in the followup PR I will type the tag arg of the function to be
`SiloedTag`,
3. in the last PR I will drop pagination from this endpoint.

Re point 3 - pagination here doesn't really make sense because we get
more logs per tag only if there are multiple devices sending logs from a
given sender to a recipient.