refactor!: dropping pagination from getLogsByTags (AztecProtocol/aztec-packages#19161)
In this PR I drop pagination from `getLogsByTags` endpoint because
pagination here doesn't really make sense because we get more logs per
tag only if there are multiple devices sending logs from a given sender
to a recipient.

This is the final of 3 PRs in which I clean up the `getLogsByTags`
endpoint:
1. First PR - including block timestamp in return value,
2. This is the second PR - I type the tag arg of the function to be
`SiloedTag`,
3. in the last PR I will drop pagination from this endpoint.

# Potential DoS vector

Log tag is unconstrained and hence can be arbitrarily chosen by the
sender. This introduces a potential DoS vector because the node
currently loads all the logs into memory and returns them. Hence it
could be feasible to re-use the same tag thousands of time and then spam
a node by request for this tag.

I think the solution here is to simply define the maximum number of logs
a node is willing to store per tag in the db. Given that there is no
legitimate use case for log reuse I think making it even as low as 5
would be fine.

Note that this PR didn't introduce this DoS vector because even though I
dropped the pagination here it didn't seem to be enforced before (if
`limitPerTag` arg was undefined we just returned it all). For this
reason I think we can merge it and tackle the issue in a followup PR.