chore: public processor should not do slow Tx.clone - second attempt (AztecProtocol/aztec-packages#19221)
As shown in https://github.com/AztecProtocol/aztec-packages/pull/19137,
tx is never mutated even by e2e tests. I think this cloning was legacy
because we _used_ to modify the Tx with contracts and effects I believe.

Second attempt after
https://github.com/AztecProtocol/aztec-packages/pull/19154 was reverted
due to flakes fixed in
https://github.com/AztecProtocol/aztec-packages/pull/19219

It turns out that `Tx.clone` was basically serving as a `sleep` that was
letting us avoid this race condition.