fix: Handle prunes when computing current base fees (AztecProtocol/aztec-packages#19234)
In [this failed run](http://ci.aztec-labs.com/b0c735d3bf1a0925) of
`e2e_cross_chain_messaging l1_to_l2` we see a
`Rollup__UnavailableTempCheckpointLog` that's coming from
`computeCurrentBaseFees`.

This was caused by trying to send a tx while a prune was happening. This
adds retries to `getPendingCheckpoint` so it tries getting the latest
checkpoint again if it fails because the checkpoint number it's trying
to fetch has been pruned out.